+++ client/src/pages/NotFound.jsx
@@ -1,0 +1,5 @@
+function NotFound() {
+  return <h1>404 - Page Not Found</h1>;
+}
+
+export default NotFound;
